Renders Image as Image instead of as link
