--- tests/Marks/ImageTest.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace HtmlToProseMirror\Test\Marks;
-
-use HtmlToProseMirror\Renderer;
-use HtmlToProseMirror\Test\TestCase;
-
-class ImageTest extends TestCase
-{
-    /** @test */
-    public function image_gets_rendered_as_link()
-    {
-        $html = '<img src="https://example.com/eggs.png" alt="The Finished Dish" />';
-
-        $json = [
-            'type' => 'doc',
-            'content' => [
-                [
-                    'type' => 'text',
-                    'text' => 'https://example.com/eggs.png',
-                    'marks' => [
-                        [
-                            'type' => 'link',
-                            'attrs' => [
-                                'href' => 'https://example.com/eggs.png',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-
-        $this->assertEquals($json, (new Renderer)->render($html));
-    }
-}
